requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
pandas>=1.2.4
Pillow>=8.2.0
nltk>=3.6.2
opencv-python>=4.5.2
matplotlib>=3.4.2
tqdm>=4.61.0
scikit-learn>=0.24.2
h5py>=3.2.1
tensorboard>=2.6.0
rouge>=1.0.0
transformers>=4.30.0 
wordcloud>=1.8.0
seaborn>=0.11.0
kagglehub>=0.1.0  